sprint 6

+++ app/controllers/serviceController.js
@@ -1,4 +1,4 @@
-app.controller('ServiceController', function($controller, $route, $scope, ProjectService, Service, ServiceRepo, NgTableParams) {
+app.controller('ServiceController', function ($controller, $scope, ProjectService, Service, ServiceRepo) {
 
     angular.extend(this, $controller('AbstractScheduleController', {
         $scope: $scope
@@ -9,7 +9,7 @@ app.controller('ServiceController', function($controller, $route, $scope, Projec
         type: "service",
         options: ['UP', 'DOWN', 'MAINTENANCE']
     };
-    
+
     $scope.serviceRepo = ServiceRepo;
 
     $scope.services = $scope.serviceRepo.getAll();
@@ -18,62 +18,65 @@ app.controller('ServiceController', function($controller, $route, $scope, Projec
 
     $scope.serviceToDelete = {};
 
-    $scope.weaverTable = {
-        repo: $scope.serviceRepo,
-        columns: [{
-                gloss: 'Service',
+    ServiceRepo.ready().then(function () {
+        $scope.weaverTable = {
+            pageSettings: $scope.serviceRepo.getPageSettings(),
+            tableParams: $scope.serviceRepo.getTableParams(),
+            columns: [{
+                    gloss: 'Service',
+                    property: 'name',
+                    filterable: true,
+                    sortable: true
+                },
+                {
+                    gloss: 'Status',
+                    property: 'status',
+                    filterable: true,
+                    sortable: true
+                },
+                {
+                    gloss: 'Auto Updating URL',
+                    property: 'serviceUrl',
+                    filterable: true,
+                    sortable: true
+                },
+                {
+                    gloss: 'Visible to Anonymous Users',
+                    property: 'isPublic',
+                    filterable: true,
+                    sortable: true
+                },
+                {
+                    gloss: 'Prominent Display',
+                    property: 'onShortList',
+                    filterable: true,
+                    sortable: true
+                },
+                {
+                    gloss: 'Project',
+                    filterable: false,
+                    sortable: false
+                },
+                {
+                    gloss: 'Actions',
+                    filterable: false,
+                    sortable: false
+                }
+            ],
+            activeSort: [{
                 property: 'name',
-                filterable: true,
-                sortable: true
-            },
-            {
-                gloss: 'Status',
-                property: 'status',
-                filterable: true,
-                sortable: true
-            },
-            {
-                gloss: 'Auto Updating URL',
-                property: 'serviceUrl',
-                filterable: true,
-                sortable: true
-            },
-            {
-                gloss: 'Public',
-                property: 'isPublic',
-                filterable: true,
-                sortable: true
-            },
-            {
-                gloss: 'Short List',
-                property: 'onShortList',
-                filterable: true,
-                sortable: true
-            },
-            {
-                gloss: 'Project',
-                filterable: false,
-                sortable: false
-            },
-            {
-                gloss: 'Actions',
-                filterable: false,
-                sortable: false
-            }
-        ],
-        activeSort: [{
-            property: 'name',
-            direction: 'ASC'
-        }]
-    };
+                direction: 'ASC'
+            }]
+        };
+    });
 
-    ProjectService.getAll().then(function(projects) {
+    ProjectService.getAll().then(function (projects) {
         $scope.projects = projects;
 
-        $scope.getProject = function(service) {
+        $scope.getProject = function (service) {
             if (service.projectId && !service.project) {
                 service.project = {};
-                ProjectService.getById(service.projectId).then(function(project) {
+                ProjectService.getById(service.projectId).then(function (project) {
                     angular.extend(service, {
                         project: project
                     });
@@ -85,7 +88,7 @@ app.controller('ServiceController', function($controller, $route, $scope, Projec
 
     });
 
-    $scope.resetServices = function() {
+    $scope.resetServices = function () {
         if ($scope.serviceData) {
             $scope.serviceData.refresh();
             $scope.serviceData.clearValidationResults();
@@ -108,59 +111,62 @@ app.controller('ServiceController', function($controller, $route, $scope, Projec
 
     $scope.resetServices();
 
-    $scope.createService = function() {
+    $scope.createService = function () {
         if ($scope.serviceData.isAuto) {
             $scope.serviceData.status = 'UP';
         } else {
             $scope.serviceData.isAuto = false;
         }
-        $scope.serviceRepo.create($scope.serviceData).then(function(res) {
+        $scope.serviceRepo.create($scope.serviceData).then(function (res) {
             if (angular.fromJson(res.body).meta.status === 'SUCCESS') {
                 $scope.resetServices();
             }
         });
     };
 
-    $scope.editService = function(service) {
+    $scope.editService = function (service) {
         $scope.serviceData = service;
         $scope.openModal('#editServiceModal');
     };
 
-    $scope.updateService = function() {
-    	$scope.serviceRepo.update($scope.serviceData).then(function(res) {
+    $scope.updateService = function () {
+        $scope.serviceData.dirty(true);
+        $scope.serviceRepo.update($scope.serviceData).then(function (res) {
             if (angular.fromJson(res.body).meta.status === 'SUCCESS') {
                 $scope.resetServices();
             }
         });
     };
 
-    $scope.editSchedule = function(service) {
+    $scope.editSchedule = function (service) {
         $scope.data = service;
         $scope.openModal('#editScheduleModal');
     };
 
-    $scope.resetSchedule = function() {
+    $scope.resetSchedule = function () {
         $scope.resetServices();
     };
 
-    ServiceRepo.ready().then(function() {
+    ServiceRepo.ready().then(function () {
         $scope.tableParams = ServiceRepo.getTableParams();
         $scope.resetServices();
     });
 
-    $scope.confirmDelete = function(service) {
+    $scope.confirmDelete = function (service) {
         $scope.openModal('#deleteServiceModal');
         $scope.serviceToDelete = service;
     };
 
-    $scope.deleteService = function() {
+    $scope.deleteService = function () {
         $scope.deleting = true;
-        $scope.serviceToDelete.delete().then(function() {
-            $scope.closeModal();
-            $scope.deleting = false;
-            ServiceRepo.remove($scope.serviceToDelete);
-            $scope.serviceToDelete = {};
-            $scope.tableParams.reload();
+        $scope.serviceToDelete.delete().then(function (res) {
+            if (angular.fromJson(res.body).meta.status === 'SUCCESS') {
+                $scope.closeModal();
+                $scope.deleting = false;
+                ServiceRepo.remove($scope.serviceToDelete);
+                $scope.serviceToDelete = {};
+                $scope.tableParams.reload();
+            }
         });
     };
